feat: move resize observer to context and refactor styles
