+++ apps/dashboard/app/(components)/new-transaction-card/styles.ts
@@ -1,3 +1,5 @@
+import { MOBILE_SIZE } from "app/context/dashboard-context.utils";
+
 const styles = {
   pixelsTop: {
     position: "absolute",
@@ -10,6 +12,16 @@ const styles = {
     zIndex: 0,
     left: 0,
     bottom: -7,
+  },
+  card: {
+    height: "420px",
+    position: "relative",
+    [`@media (max-width: ${MOBILE_SIZE}px)`]: {
+      height: "570px",
+      display: "flex",
+      flexDirection: "column",
+      alignItems: "center",
+    },
   },
 };
